Extract workspace main pane view
